split up earnings data on artemis so that formatting is consistent
