+++ game/js/src/containers/artemis/components/actualearnings/index.js
@@ -13,7 +13,7 @@ import {
   useStockOverview,
 } from '@Components/reducers/stockoverview';
 
-export const EarningsEstimatesTable = (props) => {
+export const ActualEarnings = (props) => {
   const {
     state,
   } = useStockOverview();
@@ -22,10 +22,26 @@ export const EarningsEstimatesTable = (props) => {
     earnings,
   } = state;
 
-  const actualYears = earnings.eps_est_dates.map(
-      (date) => date.substring(0, 4));
-  const actualDayMonth = earnings.eps_est_dates.map(
-      (date) => date.substring(5));
+  if (earnings == null) {
+    return null;
+  }
+
+  /* Actual EPS date contains some values for next q and next year,
+  so we need to drop nulls. Way below is N rather than N*number of
+  rows. Should be 2 or 3 nulls but this way is guaranteed.
+  */
+  const actualEpsLen = earnings.eps_actual.length;
+  const actualEpsValueLen = earnings.eps_actual.filter((v) => v!= -1).length;
+  const startIndex = actualEpsLen - actualEpsValueLen;
+  const endIndex = actualEpsLen - 1;
+
+  const actualDates = earnings.eps_dates_period.slice(startIndex, endIndex);
+  const actualEps = earnings.eps_actual.slice(startIndex, endIndex);
+  const actualEst = earnings.eps_est.slice(startIndex, endIndex);
+  const actualDiff = earnings.eps_diff.slice(startIndex, endIndex);
+
+  const actualYears = actualDates.map((date) => date.substring(0, 4));
+  const actualDayMonth = actualDates.map((date) => date.substring(5));
   const dateTitles = [
     'Year',
     'Date',
@@ -44,24 +60,16 @@ export const EarningsEstimatesTable = (props) => {
     );
   });
 
-  const bodyData = [
-    earnings.eps_est_period,
-    earnings.eps_est_avg,
-    earnings.eps_est_trend_thirty,
-    earnings.eps_est_trend_ninety,
-    earnings.eps_est_year_ago,
-    earnings.eps_est_analyst_no,
-  ];
-
-  const bodyTitles = [
-    'Period',
+  const bodyTitles= [
+    'Act',
     'Est',
-    '30 day -',
-    '90 day -',
-    '1 yr -',
-    '#',
+    'Diff',
   ];
-  const bodyRows = bodyData.map((series, i) => {
+  const bodyRows = [
+    actualEps,
+    actualEst,
+    actualDiff,
+  ].map((series, i) => {
     const title = bodyTitles[i];
     return (
       <React.Fragment
@@ -99,7 +107,7 @@ export const EarningsEstimatesTable = (props) => {
     <>
       <Title
         light>
-        Earnings Estimates
+        Actual Earnings
       </Title>
       <DefaultHorizontalSpacer>
         <ScrollableTable
